Move `MinimalScoutReportInfo` to file for its generating endpoint
