+++ lib/reusable/lovat_api/get_scout_reports_by_long_match_key.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
+import 'package:frc_8033_scouting_shared/frc_8033_scouting_shared.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/lovat_api.dart';
 
 extension GetScoutReportsByLongMatchKey on LovatAPI {
@@ -17,5 +18,28 @@ extension GetScoutReportsByLongMatchKey on LovatAPI {
     final json = jsonDecode(response!.body) as List<dynamic>;
 
     return json.map((e) => MinimalScoutReportInfo.fromJson(e)).toList();
+  }
+}
+
+class MinimalScoutReportInfo {
+  const MinimalScoutReportInfo({
+    required this.uuid,
+    required this.scout,
+    required this.timestamp,
+  });
+
+  final String uuid;
+  final Scout scout;
+  final DateTime timestamp;
+
+  factory MinimalScoutReportInfo.fromJson(Map<String, dynamic> json) {
+    return MinimalScoutReportInfo(
+      uuid: json['uuid'],
+      scout: Scout(
+        id: json['scouterUuid'],
+        name: json['scouter']['name'],
+      ),
+      timestamp: DateTime.parse(json['startTime']),
+    );
   }
 }
